Changed mean so that Doubles get overhanded

+++ src/src/Foldchange.java
@@ -5,7 +5,7 @@ public class Foldchange {
 
     public static List<List<Double>> estimateFoldchange(Counts counts){
         List<List<Double>> foldchange = new ArrayList<>();
-        List<List<Integer>> mean = calculateRowStatistics.meanCalculator(counts);
+        List<List<Double>> mean = calculateRowStatistics.meanCalculator(counts);
 
         for (int i = 0; i < mean.size(); i++){
             double mean0 = mean.get(i).get(0);
